Rename directory tutorings to tutorial
